+++ gateway/upstreamer/push/notifier.go
@@ -1,4 +1,4 @@
-package pushupstreamer
+package push
 
 import (
 	"context"
