+++ gateway/upstreamer/push/options_test.go
@@ -1,4 +1,4 @@
-package pushupstreamer
+package push
 
 import (
 	"testing"
