+++ gateway/upstreamer/push/types.go
@@ -1,4 +1,4 @@
-package pushupstreamer
+package push
 
 import (
 	"go.aporeto.io/bahamut"
